% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirq.R
\name{print.dirq}
\alias{print.dirq}
\title{Print method for \code{dirq} objects}
\usage{
\method{print}{dirq}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"dirq"}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays basic information about the fitted directional quantile boundary.
}
