% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_pcor.R
\name{print.median_pcor}
\alias{print.median_pcor}
\title{Print method for \code{median_pcor} objects}
\usage{
\method{print}{median_pcor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{median_pcor}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for \code{median_pcor} objects
}
